Generalize handling of multi-value fields
Multi-value fields such as hit.heading and hit.category will have
to be handled by pretty much everything. Generalize that out
so it can be done repeatably and without errors.

Might make some aggregation and double aggregations features harder
to implement, but for now we don't have any so run with it.