Make feature_vector_space.py work again
The data sizes involved with duplicating the term vector data across all
the rows was just too much, so this has be re-done to store separately
with python shelve. Data is loaded from the shelve on-demand which is
certainly slower than in-memory, but in-memory is simply untractable
with a few hundred thousand search sessions.