Convert es doc fields storage to shelve
When dealing with large numbers of sessions we use way too much memory
duplicating the same fields across all the rows. Apply the same technique
of utilizing shelve to store and retrieve doc fields as the last patch.

This is slower, we have to unpickle the stored values and that adds 20-30% to
the runtime, but it seems the only reasonable way to handle things. Could try
pickling a giant dict, but doc fields on their own are >5G on a 10k query
sample.

Refactors feature_base.py so that these fields magic themselves into place,
rather than needing to handle it in each feature file.